feat: relocate constant for MAX_MESSAGE_SIZE
