+++ p2p/src/main/kotlin/maru/p2p/topics/TopicHandlerWithInOrderDelivering.kt
@@ -11,11 +11,11 @@ package maru.p2p.topics
 import java.util.Optional
 import java.util.PriorityQueue
 import maru.p2p.LINEA_DOMAIN
-import maru.p2p.MAX_MESSAGE_SIZE
 import maru.p2p.MaruPreparedGossipMessage
 import maru.p2p.SubscriptionManager
 import maru.p2p.ValidationResultCode
 import maru.serialization.Deserializer
+import maru.serialization.MAX_MESSAGE_SIZE
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.apache.tuweni.bytes.Bytes
